pyproject.toml
[tool.poetry]
name = "mediumroast_py"
version = "0.3.0"
description = "A package to perform ETL (Extract, Transform and Load) and generally interact with the mediumroast.io application."
authors = ["Michael Hay <michael.hay@mediumroast.io>"]
license = "Apache-2.0"
repository = "https://github.com/mediumroast/mr_python"
readme = "README.md"
keywords = ["Product Management", "Competitive Insights", "Customer Insights", "Partner Insights", "Empowered Teams"]
include = ["examples/*", "examples/*/*", "examples/.*/*"]


[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
pyfiglet = "^0.8.post1"
spacy = "^3.3.1"
boto3 = "^1.24.24"
python-magic = "^0.4.27"
pdfx = "^1.4.1"
PyDocX = "^0.9.10"
python-pptx = "^0.6.21"
geopy = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/mediumroast/mr_python/issues"
